Get rid of various 64bit build warning: size_t/int casts, and DLL interfaces for internal classes put into a QValueList.
[git-p4: depot-paths = "//depot/qt/main/": change = 47721]

+++ src/kernel/qobject.cpp
@@ -215,7 +215,7 @@ static QCString qt_rmWS( const char *s )
     result.truncate( (int)(d - result.data()) );
     int void_pos = result.find("(void)");
     if ( void_pos >= 0 )
-	result.remove( void_pos+1, strlen("void") );
+	result.remove( void_pos+1, (uint)strlen("void") );
     return result;
 }
